% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getFitnessEvolution,GenAlg,missing-method}
\alias{fitness}
\alias{fitness,GenAlg-method}
\alias{getFitnessEvolution,GenAlg,character-method}
\alias{getFitnessEvolution,GenAlg,missing-method}
\title{Get the fitness of a variable subset}
\usage{
\S4method{getFitnessEvolution}{GenAlg,missing}(object, type = "avg")

\S4method{getFitnessEvolution}{GenAlg,character}(object, type = "avg")

fitness(object)

\S4method{fitness}{GenAlg}(object)
}
\arguments{
\item{object}{The \code{\link{GenAlg}} object returned by \code{\link{genAlg}}}

\item{type}{can be \code{"best"} to return the fitness of the best chromosome for each generation
or \code{"avg"} to return the average fitness during each generation}
}
\value{
A vector with the estimated fitness for each solution
}
\description{
Get the internal fitness for all variable subsets
}
\details{
This method is used to get the fitness of all variable subsets
found by the genetic algorithm.
}
\examples{
ctrl <- genAlgControl(populationSize = 200, numGenerations = 30,
	minVariables = 5, maxVariables = 15)
evaluator <- evaluatorPLS(numReplications = 3L, innerSegments = 4L, numThreads = 2L)

\dontrun{
result <- genAlg(y, x, control = ctrl, evaluator = evaluator)

fitness(result)
}
}

