\docType{methods}
\name{validData}
\alias{validData}
\alias{validData,GenAlgEvaluator,GenAlg-method}
\alias{validData,GenAlgLMEvaluator,GenAlg-method}
\alias{validData,GenAlgPLSEvaluator,GenAlg-method}
\title{Check if the data is valid for the evaluator}
\arguments{
  \item{object}{The evaluator}

  \item{genAlg}{The GenAlg object the evaluator is used in}
}
\description{
  This method checks if the covariates matrix is valid for
  the evaluator
}

