% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{GenAlgFitEvaluator-class}
\alias{GenAlgFitEvaluator}
\alias{GenAlgFitEvaluator-class}
\title{Fit Evaluator}
\description{
Fit Evaluator
}
\section{Slots}{

\describe{
\item{\code{numSegments}}{The number of CV segments used in one replication.}

\item{\code{numThreads}}{The maximum number of threads the algorithm is allowed to spawn (a value less than 1 or NULL means no threads).}

\item{\code{statistic}}{The statistic used to evaluate the fitness.}

\item{\code{statisticId}}{The (internal) numeric ID of the statistic.}
}}

