% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluator.R
\docType{class}
\name{GenAlgLMEvaluator-class}
\alias{GenAlgLMEvaluator-class}
\alias{GenAlgLMEvaluator}
\title{LM Evaluator}
\description{
LM Evaluator
}
\section{Slots}{

\describe{
\item{\code{statistic}}{The statistic used to evaluate the fitness.}

\item{\code{statisticId}}{The (internal) numeric ID of the statistic.}

\item{\code{numThreads}}{The maximum number of threads the algorithm is allowed to spawn (a value less than 1 or NULL means no threads).}
}}

