% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluator.R
\docType{class}
\name{GenAlgUserEvaluator-class}
\alias{GenAlgUserEvaluator-class}
\alias{GenAlgUserEvaluator}
\title{User Function Evaluator}
\description{
User Function Evaluator
}
\section{Slots}{

\describe{
\item{\code{evalFunction}}{The function that is called to evaluate the variable subset.}

\item{\code{sepFunction}}{The function that calculates the standard error of prediction for the found subsets.}
}}

