% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getFitnessEvolution}
\alias{getFitnessEvolution}
\title{Get the evolution of the fitness}
\usage{
getFitnessEvolution(object, type = c("avg", "best"))
}
\arguments{
\item{object}{The \code{\link{GenAlg}} object returned by \code{\link{genAlg}}}

\item{type}{can be \code{"best"} to return the fitness of the best chromosome for each generation
or \code{"avg"} to return the average fitness during each generation}
}
\value{
A vector with the best or average fitness value after each generation
}
\description{
Get the fitness of the best / average chromosomes after each generation
}
\details{
Returns the progress of the fitness of the best or average chromosome.
}

