\docType{methods}
\name{SEPeval}
\alias{SEPeval}
\alias{SEPeval,GenAlgPLSEvaluator,GenAlg-method}
\alias{SEPeval,GenAlgUserEvaluator,GenAlg-method}
\title{Get the SEP of all variable subsets}
\arguments{
  \item{object}{The used evaluator}

  \item{genAlg}{The GenAlg object}
}
\description{
  Calculate the standard error of prediction for all
  variable subsets
}
\details{
  This method is used to get the standard error of
  prediction (SEP) of all variable subsets found by the
  genetic algorithm.
}

