% Generated by roxygen2 (4.0.1): do not edit by hand
\name{subsets}
\alias{subsets}
\title{Get the found variable subset(s)}
\usage{
subsets(object, indices, names = TRUE)
}
\arguments{
\item{object}{The GenAlg object returned by \code{\link{genAlg}}.}

\item{indices}{The indices of the subsets or empty if all subsets should be returned.}

\item{names}{Should the names or the column numbers of the variables be returned.}
}
\value{
A logical matrix where each column represents a variable subset
}
\description{
Get a list of variable indices/names of the found variable subsets.
}
\details{
This method is used to get the names or indices of the variables used in specified variable subsets.
}
\examples{
ctrl <- genAlgControl(populationSize = 200, numGenerations = 30,
	minVariables = 5, maxVariables = 15)
evaluator <- evaluatorPLS(numReplications = 3L, innerSegments = 4L, numThreads = 2L)

\dontrun{
result <- genAlg(y, x, control = ctrl, evaluator = evaluator)
subsets(result, names = TRUE, indices = 1:5) # best 5 variable subsets as a list of names
result@subsets[ , 1:5] # best 5 variable subsets as a logical matrix with the subsets in the columns
}
}

