\docType{methods}
\name{toCControlList}
\alias{toCControlList}
\alias{toCControlList,GenAlgControl-method}
\alias{toCControlList,GenAlgLMEvaluator-method}
\alias{toCControlList,GenAlgPLSEvaluator-method}
\alias{toCControlList,GenAlgUserEvaluator-method}
\title{Transform the object to a list}
\arguments{
  \item{object}{The object}
}
\description{
  Get the control list for the C++ procedure genAlgPLS from
  the object
}

