% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\docType{methods}
\name{trueFitnessVal}
\alias{trueFitnessVal}
\alias{trueFitnessVal,GenAlgPLSEvaluator,numeric-method}
\alias{trueFitnessVal,GenAlgUserEvaluator,numeric-method}
\alias{trueFitnessVal,GenAlgLMEvaluator,numeric-method}
\alias{trueFitnessVal,GenAlgFitEvaluator,numeric-method}
\title{Get the transformed fitness values}
\usage{
trueFitnessVal(object, fitness)

\S4method{trueFitnessVal}{GenAlgPLSEvaluator,numeric}(object, fitness)

\S4method{trueFitnessVal}{GenAlgUserEvaluator,numeric}(object, fitness)

\S4method{trueFitnessVal}{GenAlgLMEvaluator,numeric}(object, fitness)

\S4method{trueFitnessVal}{GenAlgFitEvaluator,numeric}(object, fitness)
}
\arguments{
\item{object}{The used evaluator, an object with type or with a subtype of \code{\link{GenAlgEvaluator}}}

\item{fitness}{A numeric vector of fitnesses}
}
\value{
A vector with the true fitness values
}
\description{
Transform the given fitness values according tho the GenAlgEvaluator class
}
\details{
This method is used to calculate the true fitness given the GenAlgEvaluator class (as they use
different internal fitness measures)
}
