% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\alias{evaluate,GenAlgEvaluator,matrix,numeric,matrix,integer,integer-method}
\alias{evaluate,GenAlgEvaluator,matrix,numeric,logical,integer,integer-method}
\alias{evaluate,GenAlgEvaluator,matrix,numeric,ANY,missing,integer-method}
\alias{evaluate,GenAlgEvaluator,matrix,numeric,ANY,integer,missing-method}
\alias{evaluate,GenAlgEvaluator,matrix,numeric,ANY,missing,missing-method}
\title{Evaluate the fitness of variable subsets}
\usage{
evaluate(object, X, y, subsets, seed, verbosity)

\S4method{evaluate}{GenAlgEvaluator,matrix,numeric,matrix,integer,integer}(object, X, y, subsets, seed, verbosity)

\S4method{evaluate}{GenAlgEvaluator,matrix,numeric,logical,integer,integer}(object, X, y, subsets, seed, verbosity)

\S4method{evaluate}{GenAlgEvaluator,matrix,numeric,ANY,missing,integer}(object, X, y, subsets, seed, verbosity)

\S4method{evaluate}{GenAlgEvaluator,matrix,numeric,ANY,integer,missing}(object, X, y, subsets, seed, verbosity)

\S4method{evaluate}{GenAlgEvaluator,matrix,numeric,ANY,missing,missing}(object, X, y, subsets, seed, verbosity)
}
\arguments{
\item{object}{The GenAlgEvaluator object that is used to evaluate the variables}

\item{X}{The data matrix used to for fitting the model}

\item{y}{The response vector}

\item{subsets}{The logical matrix where a column stands for one subset to evaluate}

\item{seed}{The value to seed the random number generator before evaluating}

\item{verbosity}{A value between 0 (no output at all) and 5 (maximum verbosity)}
}
\description{
Evaluate the given variable subsets with the given Evaluator
}
