\docType{methods}
\name{evaluate}
\alias{evaluate}
\alias{evaluate,ANY,matrix,numeric,NULL-method}
\alias{evaluate,ANY,matrix,numeric,numeric-method}
\alias{evaluate,GenAlgEvaluator,matrix,numeric,integer-method}
\alias{evaluate,GenAlgUserEvaluator,matrix,numeric,integer-method}
\title{Evaluate the fitness of variables}
\arguments{
  \item{object}{The GenAlgEvaluator object that is used to
  evaluate the variables}
}
\description{
  Evaluate the given data with the given evaluator
}

