% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fitnessEvolution}
\alias{fitnessEvolution}
\title{Get the evolution of the fitness}
\usage{
fitnessEvolution(object, type = c("mean", "best", "std.dev"))
}
\arguments{
\item{object}{The \code{\link{GenAlg}} object returned by \code{\link{genAlg}}}

\item{type}{can be one ore more of \code{"best"} (to return the fitness of the best chromosome for each generation),
\code{"mean"} (to return the arithmetic mean fitness during each generation), and \code{"std.dev"} (for
the standard deviation of the fitness values in each generation).}
}
\value{
A vector with the best or average fitness value after each generation
}
\description{
Get the fitness of the best / average chromosomes after each generation
}
\details{
Returns the progress of the fitness of the best or average chromosome.
}

