\name{GenAlgLMEvaluator}
\alias{GenAlgLMEvaluator}
\title{LM Evaluator}
\description{
  LM Evaluator
}
\section{Slots}{
  \describe{ \item{\code{statistic}:}{The statistic used to
  evaluate the fitness.} \item{\code{maxCor}:}{If a
  correlation between the covariates in a subset is higher
  than this value, the principal components of the
  covariates are used to fit the linear model.} }
}

