% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{GenAlgLMEvaluator-class}
\alias{GenAlgLMEvaluator}
\alias{GenAlgLMEvaluator-class}
\title{LM Evaluator}
\description{
LM Evaluator
}
\section{Slots}{

\describe{
\item{\code{statistic}}{The statistic used to evaluate the fitness.}

\item{\code{statisticId}}{The (internal) numeric ID of the statistic}

\item{\code{numThreads}}{The maximum number of threads the algorithm is allowed to spawn (a value less than 1 or NULL means no threads).}
}}

