% Generated by roxygen2 (4.0.0): do not edit by hand
\name{evaluatorLM}
\alias{evaluatorLM}
\title{LM Evaluator}
\usage{
evaluatorLM(statistic = c("BIC", "AIC", "adjusted.r.squared", "r.squared"),
  numThreads = NULL, maxCor = NULL)
}
\arguments{
\item{statistic}{The statistic used to evaluate the fitness}

\item{numThreads}{The maximum number of threads the algorithm is allowed to spawn (a value less than 1 or NULL means no threads)}

\item{maxCor}{If the correlation-matrix of the covariates has an entry (absolutely) greater than this value
the principal components are used to calculate the fit}
}
\value{
Returns an S4 object of type \code{\link{GenAlgLMEvaluator}}
}
\description{
Create an evaluator that uses a linear model to evaluate the fitness.
}
\details{
Different statistics to evaluate the fitness of the variable subset can be given. If a maximum
absolute correlation is given the algorithm will be very slow (as the C++ implementation can not
be used anymore) and multithreading is not available.
}
\examples{
ctrl <- genAlgControl(populationSize = 200, numGenerations = 30,
	minVariables = 5, maxVariables = 15)
evaluator <- evaluatorLM(statistic = "adjusted.r.squared", maxCor = 0.8)

\dontrun{
result <- genAlg(y, x, control = ctrl, evaluator = evaluator)
}
}
\seealso{
Other GenAlg.Evaluators: \code{\link{evaluatorPLS}};
  \code{\link{evaluatorUserFunction}}
}

