% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluator.R
\docType{class}
\name{GenAlgPLSEvaluator-class}
\alias{GenAlgPLSEvaluator-class}
\alias{GenAlgPLSEvaluator}
\title{PLS Evaluator}
\description{
PLS Evaluator
}
\section{Slots}{

\describe{
\item{\code{numReplications}}{The number of replications used to evaluate a variable subset.}

\item{\code{innerSegments}}{The number of inner RDCV segments used in one replication.}

\item{\code{outerSegments}}{The number of outer RDCV segments used in one replication.}

\item{\code{testSetSize}}{The relative size of the test set (between 0 and 1).}

\item{\code{sdfact}}{The factor to scale the stand. dev. of the MSEP values when selecting the optimal number
of components. For the "one standard error rule", \code{sdfact} is 1.}

\item{\code{numThreads}}{The maximum number of threads the algorithm is allowed to spawn (a value less than 1 or NULL means no threads).}

\item{\code{maxNComp}}{The maximum number of components to consider in the PLS model.}

\item{\code{method}}{The PLS method used to fit the PLS model (currently only SIMPLS is implemented).}

\item{\code{methodId}}{The ID of the PLS method used to fit the PLS model (see C++ code for allowed values).}
}}

