% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toCControlList.R
\docType{methods}
\name{toCControlList}
\alias{toCControlList}
\alias{toCControlList,GenAlgPLSEvaluator-method}
\alias{toCControlList,GenAlgFitEvaluator-method}
\alias{toCControlList,GenAlgUserEvaluator-method}
\alias{toCControlList,GenAlgLMEvaluator-method}
\alias{toCControlList,GenAlgControl-method}
\title{Transform the object to a list}
\usage{
toCControlList(object)

\S4method{toCControlList}{GenAlgPLSEvaluator}(object)

\S4method{toCControlList}{GenAlgFitEvaluator}(object)

\S4method{toCControlList}{GenAlgUserEvaluator}(object)

\S4method{toCControlList}{GenAlgLMEvaluator}(object)

\S4method{toCControlList}{GenAlgControl}(object)
}
\arguments{
\item{object}{The object}
}
\value{
A list with all items expected by the C++ code
}
\description{
Get the control list for the C++ procedure genAlgPLS from the object
}
